pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools_scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "daylily-ursa"
dynamic = ["version"]
description = "Daylily Ursa beta analysis API"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Daylily Informatics", email = "daylily@daylilyinformatics.com"}
]
dependencies = [
    "boto3>=1.26.0",
    "boto3-stubs[s3,sns,cloudwatch]>=1.28.0",
    "bandit[toml]>=1.8.0",
    "black>=23.0.0",
    "pyyaml>=6.0",
    "sqlalchemy>=2.0.0",
    "cli-core-yo==2.1.1",
    # Cognito auth library
    "daylily-auth-cognito==2.1.5",
    # TapDB graph persistence
    "daylily-tapdb==6.0.8",
    # Ephemeral cluster API/CLI contract
    "daylily-ephemeral-cluster==2.1.11",
    # Shared Zebra runtime contract
    "zebra_day==6.0.1",
    "fastapi>=0.104.0",
    "httpx>=0.25.0",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.0",
    "jsonschema>=4.17.0",
    "moto>=4.2.0",
    "mypy>=1.5.0",
    "passlib[bcrypt]>=1.7.4",
    "playwright>=1.42.0",
    "pre-commit>=3.8.0",
    "pydantic>=2.0.0",
    "pydantic[email]>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-playwright>=0.4.4",
    "tabulate",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "ruff>=0.1.0",
    "uvicorn[standard]>=0.24.0",
]

[project.scripts]
ursa = "daylib_ursa.cli:main"
daylily-workset-api = "daylib_ursa.workset_api_cli:main"

[tool.setuptools_scm]
version_scheme = "guess-next-dev"
local_scheme = "node-and-date"

[tool.setuptools.packages.find]
include = ["daylib_ursa*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v"
markers = [
    "e2e: marks tests as end-to-end tests requiring playwright",
]

[tool.black]
line-length = 100
target-version = ['py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
no_site_packages = true

[tool.bandit]
skips = ["B101", "B104", "B105", "B106", "B108", "B112", "B404", "B603", "B607"]
